Draft: proposal for unit & integration testing
There's a lot going on, and I haven't documented any of it. But the idea
is to be able to do unit tests without relying on real git, since it
could be slow, use the network, or impact local state.

So `cargo test --lib` will use mostly in-memory tests, reaching out to
fake git binaries when necessary. The Integration tests (under the
`tests/` folder) will run against the real git binary, and may do things
like authenticate to remote hosts or create / remove branches.