Use 'debug' or 'release' bins as appropriate.
I kept accidentally running `cargo test` which would fail because the
fake binaries hadn't been updated; this is because we hardcoded the
"release" binaries since our CI policy is to test with the "--release"
flag, and my invocation of `cargo test` would only update the "debug"
binaries.

The solution was either to develop `cargo test --release` as a reflex,
or write a macro to ensure we're using the correct (and therefore most
recently updated) set of mocks no matter which test command we use.