DRAFT: Create a git repo for each test
This is a proof of concept for re-introducing the `tempdir` package and
using it to create a git repository for each test. By adding a
`working_dir` field to the Git client wrapper, it can run all of its git
commands with the `-C` flag, as in:

```sh
git -C /tmp/dir/git-pr-test-12345 branch -va
```

According to [GIT(1)][1], the `-C <DIR>` option instructs git to run as
though it had been launched from `DIR`, even if `DIR` is not the current
working directory of the parent process. Given that we cannot change the
working directory of `cargo test` (since it runs tests as threads of a
single process, and the notion of a "working directory" is a
process-level property), this should give us exactly what we want.

**Caveat.** The temporary repositories do not seem to actually get deleted
on my mac. The [TempDir docs][2] seem to anticipate this:

> Various platform-specific conditions may cause TempDir to fail to
> delete the underlying directory.

So, perhaps this is all for naught. 🤷

[1]: https://www.man7.org/linux/man-pages/man1/git.1.html
[2]: https://docs.rs/tempdir/0.3.7/tempdir/struct.TempDir.html