Add UNIX manual entry. Fixes #29.
This commit implicitly proposes that we combine all of the `git-pr-X`
commands into `git-pr --X`. After reviewing the way git's own
subcommands are organized, I believe this is the more appropriate style.
We will then be able to have a single [clap][1] instance from which we
can generate tab-completion modules.

[1]: https://clap.rs